Mark expressions nullable by grouping sets.
When generating window_pathkeys, distinct_pathkeys or sort_pathkeys, we
failed to realize that the grouping/ordering expressions might be
nullable by grouping sets.  As a result, we may incorrectly deem that
the PathKeys are redundant by EquivalenceClass processing and thus
remove them from the pathkeys list.  That would lead to wrong results in
some cases.

To fix this issue, we mark the grouped expressions nullable by grouping
sets if that is the case.  If the grouped expression is a Var or
PlaceHolderVar or constructed from those, we can just add the RT index
of the GROUP RTE to the existing nullingrels field(s); otherwise we have
to add a PlaceHolderVar to carry on the nullingrel bit.

However, we have to manually remove this nullingrel bit from expressions
in various cases where these expressions are logically below the
grouping step, such as when we generate groupClause pathkeys for
grouping sets, or when we generate PathTarget for initial input to
grouping nodes.

Furthermore, in set_upper_references, the targetlist and quals of an Agg
node should have nullingrels that include the effects of the grouping
step, ie they will have nullingrels equal to the input Vars/PHVs'
nullingrels plus the nullingrel bit that references the grouping RTE.
In order to perform exact nullingrels matches, we also need to manually
remove this nullingrel bit.