Introduce an RTE for the grouping step
If there are subqueries in the grouping expressions, each of these
subqueries in the targetlist and HAVING clause is expanded into distinct
SubPlan nodes.  As a result, most of these SubPlan nodes cannot get
converted to references to the grouping key column output by the Agg
node, which means that they have to get evaluated afresh.  This causes
problems with grouping sets in cases where they should go to NULL
because they are from the wrong grouping set.  Furthermore, during
re-evaluation, these SubPlan nodes might use the nulled value of some
columns, which is not correct.

To fix this issue, we introduce a new kind of RTE representing the
output of the grouping step, with columns that are the Vars or
expressions being grouped on.  In the parser, we replace the grouping
expressions in the targetlist and HAVING clause with Vars referencing
this new RTE, so that the output of the parser directly expresses the
semantic requirement that certain grouping expressions be gotten from
the grouping output rather than computed some other way.  In the
planner, we first preprocess all the columns of this new RTE and then
replace any Vars in the targetlist and HAVING clause that reference this
new RTE with the underlying grouping expressions, so that we will have
only one instance of a SubPlan node for each subquery contained in the
grouping expressions.