Check lateral refs within PHVs for memoize cache keys
If we intend to generate a memoize node on top of a path, we need cache
keys of some sort.  Currently we search for the cache keys in the
parameterized clauses of the path as well as the lateral_vars of its
parent.  However, it turns out that this is not sufficient because their
might be lateral references derived from PlaceHolderVars, which we fail
to take into consideration.

This oversight can cause us to miss opportunities to utilize the Memoize
node.  Moreover, in some plans, the failure to recognize all the cache
keys could result in performance regressions.  This is because without
identifying all the cache keys, we would need to purge the entire cache
every time we get a new outer tuple during execution.

This patch fixes this issue by extracting lateral Vars from within
PlaceHolderVars and subsequently including them in the cache keys.

This patch also includes a comment clarifying that Memoize nodes are
currently not added on top of join relation paths.  This explains why
this patch only considers PHVs that are due to be evaluated at baserels.