Remove no-op PlaceHolderVars
We may insert PlaceHolderVars when pulling up a subquery that is
within the nullable side of an outer join.  However, if the outer join
is later reduced to an inner join, the PHVs would become no-ops with
no phnullingrels bits.

It's always desirable to remove these no-op PlaceHolderVars because
they can constrain optimization opportunities, such as blocking
subexpression folding, or forcing join order.  There is one case where
we need to keep a PHV even if its phnullingrels becomes empty: the PHV
is used to enforce separate identity of subexpressions.

This patch removes no-op PlaceHolderVars by marking PHVs that need to
be kept because they are serving to isolate subexpressions.