Fix pushdown of degenerate HAVING clauses
Commit 67a54b9e8 taught the planner to push down HAVING clauses even
when nonempty grouping sets are present, as long as the clause does
not reference any columns that are nullable by the grouping sets.
However, there was an oversight: if any empty grouping sets are also
present, the aggregation node can produce a row that did not come from
the input, and pushing down a HAVING clause in this case may cause us
to fail to filter out that row.

Currently, non-degenerate HAVING clauses are not pushed down when
empty grouping sets are present, since the empty grouping sets would
nullify the vars they reference.  However, degenerate (variable-free)
HAVING clauses are not subject to this restriction and may be
incorrectly pushed down.

To fix, this patch adds a check that prevents degenerate HAVING
clauses from being pushed down when any nonempty grouping sets are
present.  This could be done more precisely, for example by
restricting the prevention only to cases where empty grouping sets are
also present, but the added complexity does not seem justified.

Back-patch to v18 where this issue was introduced.