Simplify COALESCE expressions using proven non-null arguments
The COALESCE function returns the first of its arguments that is not
null.  When an argument is proven non-null, if it is the first
non-null-constant argument, the entire COALESCE expression can be
replaced by that argument.  If it is a subsequent argument, all
following arguments can be dropped, since they will never be reached.

Currently, we apply this simplification only to Const arguments.  We
can extend this logic to Var arguments, since we now have the NOT NULL
attribute information available during constant folding.  We can also
extend it to CoalesceExpr arguments based on the knowledge that a
CoalesceExpr cannot be NULL if at least one of its arguments can be
proven non-nullable.

This can help avoid the overhead of evaluating unreachable arguments.
It can also lead to better plans when the first argument is a NOT NULL
column and thus replaces the expression, as the planner no longer has
to treat the expression as non-strict, and can also leverage index
scans on that column.

Additionally, if we've determined that a CoalesceExpr is non-nullable,
we can reduce its corresponding NullTest quals to a constant during
constant folding.  It might be argued that it is uncommon to write
"COALESCE(...) IS [NOT] NULL" by hand, but this pattern is likely not
uncommon after view expansion, function inlining, or ORM query
generation.  Also, since the reduction does not cost much, I believe
the benefit justifies the cost.